refactor: grouped hooks by service
