+++ src/components/shared/navbar/navbar.tsx
@@ -1,5 +1,5 @@
 "use client"
-import { useCurrentUserImage } from "@/utils/hooks/useCurrentUserImage";
+import { useCurrentUserImage } from "@/utils/hooks/user/useCurrentUserImage";
 import React, { useState } from "react";
 import Image from "next/image";
 import { User, Menu, X, Home, Settings, LogOut, Orbit } from "lucide-react";
